[nextest-runner] move reporter events into their own file
These events are shared by all reporter code, so it makes sense to put them in
their own file.
